pyproject.toml
[project]
name = "fitnessplan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.1",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "reportlab>=4.4.4",
    "seaborn>=0.13.2",
]
